Remove metaspace limits for now
We used to test Gradle with a permgen limit of 320Mb.
However, trying to enforce the same limit for metaspace
on Java8+ consistently leads to OutOfMemoryErrors.

Even raising it significantly to 1G does not seem to solve
all these issues. The only explanation I currently have is
that we introduced a leak between now and the point when we
stopped testing for Java 7.